Fixing linting errors
